requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
chronology==1.1
click==8.1.8
cloudpickle==3.1.1
contourpy==1.3.1
cycler==0.12.1
dacite==1.9.2
distro==1.9.0
entrypoints==0.4
fake-useragent==2.0.3
Faker==36.1.1
favicon==0.7.0
fonttools==4.56.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
htbuilder==0.9.0
htmlmin==0.1.12
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ImageHash==4.3.1
imageio==2.37.0
jaro-winkler==2.0.3
jellyfish==1.1.3
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
kmodes==0.12.2
lazy_loader==0.4
llvmlite==0.44.0
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
missingno==0.5.2
multimethod==1.12
narwhals==1.28.0
navigation==0.1
navigation-utilities==0.2.2
neo4j==5.28.1
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai==1.65.3
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
patsy==1.0.1
phik==0.12.4
pillow==11.1.0
plotly==6.0.0
pluralizer==1.2.0
prometheus_client==0.21.1
protobuf==5.29.3
pyarrow==19.0.1
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pymdown-extensions==10.14.3
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyWavelets==1.8.0
PyYAML==6.0.2
recordlinkage==0.16
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
scikit-fuzzy==0.5.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sections==0.0.3
selenium==4.29.0
six==1.17.0
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
st-annotated-text==4.0.2
st-theme==1.2.3
statsmodels==0.14.4
streamlit==1.42.2
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.5.5
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.3.0
streamlit-nested-layout==0.1.4
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
streamlit-ydata-profiling==0.2.1
tenacity==9.0.0
threadpoolctl==3.5.0
tifffile==2025.2.18
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typeguard==4.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
validators==0.34.0
visions==0.7.6
watchdog==3.0.0
websocket-client==1.8.0
wordcloud==1.9.4
wsproto==1.2.0
ydata-profiling==4.12.2
